Move all server files in server folder
