+++ test/server.test.js
@@ -1,6 +1,6 @@
 var assert = require('assert');
 
-const app = require('../app')
+const app = require('../server/app')
 const supertest = require('supertest')
 const request = supertest(app)
